requirements.txt
anyio==2.0.2
apturl==0.5.2
arabic-reshaper==2.1.1
argon2-cffi==20.1.0
asgiref==3.2.10
astroid==2.4.2
async-generator==1.10
attrs==20.3.0
Babel==2.9.0
backcall==0.2.0
bcrypt==3.1.7
bleach==3.2.2
blinker==1.4
Brlapi==0.7.0
cachetools==2.1.0
cairocffi==1.2.0
CairoSVG==2.5.1
certifi==2020.6.20
cffi==1.14.4
chardet==3.0.4
Click==7.0
colorama==0.4.3
command-not-found==0.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
cssselect2==0.4.1
cupshelpers==1.0
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
defusedxml==0.6.0
distro==1.4.0
distro-info===0.23ubuntu1
Django==2.1.5
django-allauth==0.42.0
django-cors-headers==3.4.0
django-rest-auth==0.9.5
django-rest-swagger==2.0.7
djangorestframework==3.11.0
djangorestframework-simplejwt==4.6.0
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
future==0.18.2
google-api-python-client==1.7.4
google-auth==1.5.1
google-auth-httplib2==0.0.3
html5lib==1.0.1
httplib2==0.11.3
idna==2.10
ipykernel==5.4.3
ipython==7.19.0
ipython-genutils==0.2.0
isort==5.7.0
itypes==1.2.0
jedi==0.18.0
Jinja2==2.11.2
joblib==1.0.0
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.11
jupyter-core==4.7.0
jupyter-server==1.2.2
jupyterlab==3.0.5
jupyterlab-pygments==0.1.2
jupyterlab-server==2.1.2
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
mccabe==0.6.1
mistune==0.8.4
monotonic==1.5
nbclassic==0.2.6
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.4.3
netifaces==0.10.4
notebook==6.2.0
numpy==1.19.5
oauth2client==4.1.3
oauthlib==3.1.0
olefile==0.46
openapi-codec==1.3.2
packaging==20.8
pandas==1.2.1
pandocfilters==1.4.3
paramiko==2.6.0
parso==0.8.1
pbr==5.5.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
prometheus-client==0.9.0
prompt-toolkit==3.0.14
protobuf==3.6.1
ptyprocess==0.7.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycairo==1.16.2
pycparser==2.20
pycups==1.9.73
Pygments==2.7.4
PyGObject==3.36.0
PyJWT==1.7.1
pylint==2.6.0
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.7
PyPDF2==1.26.0
Pyphen==0.10.0
pyRFC3339==1.1
pyrsistent==0.17.3
python-apt==2.0.0+ubuntu0.20.4.3
python-bidi==0.4.2
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python3-openid==3.2.0
pytz==2018.5
pyxdg==0.26
PyYAML==5.3.1
pyzmq==21.0.2
reportlab==3.5.34
requests==2.24.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
rsa==3.4.2
scikit-learn==0.24.1
scipy==1.6.0
SecretStorage==2.3.1
Send2Trash==1.5.0
simplejson==3.16.0
six==1.11.0
sniffio==1.2.0
sqlparse==0.3.1
systemd-python==234
terminado==0.9.2
testpath==0.4.4
testresources==2.0.1
threadpoolctl==2.1.0
tinycss2==1.1.0
toml==0.10.2
tornado==6.1
traitlets==5.0.5
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uritemplate==3.0.0
urllib3==1.25.10
usb-creator==0.3.7
wadllib==1.3.3
wcwidth==0.2.5
WeasyPrint==52.2
webencodings==0.5.1
wrapt==1.12.1
xhtml2pdf==0.2.5
xkit==0.0.0
